Localized Hole Scale Factor Implemented
Can compile. Have not run yet to confirm, but this is very exciting. Currently the localized scale factor goes in progressively larger n-dimensional spheres around a space, but this has potential for skewing results on the edges of dead zones or bin ranges. Something to keep in mind for systematics